requirements.txt
streamlit
streamlit-webrtc
av
tensorflow-cpu
opencv-python-headless
numpy
Pillow